pyproject.toml
[project]
name = "notime-workspace"
version = "0.1.0"
description = "A workspace for the notime project, containing shared development dependencies."
dependencies = [
    "aiogram>=3.23.0",
    "celery>=5.6.0",
    "python-dotenv>=1.2.1",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "redis>=7.1.0",
    "uvicorn>=0.38.0",
]

[project.optional-dependencies]
dev = [
    "ruff==0.14.9",
]
